#112 TempUserAppointment Monster Class 문제 해결 - TempUser 관련 Controller, Service 생성
