+++ src/main/java/com/example/gatherplan/controller/TempAppointmentController.java
@@ -175,56 +175,4 @@ public class TempAppointmentController {
                 BooleanResp.success()
         );
     }
-
-    @GetMapping("/host:check")
-    @Operation(summary = "비회원의 호스트의 여부 조회 요청", description = "비회원의 호스트 여부를 판단할 때 사용됩니다. [figma #19,#27]")
-    public ResponseEntity<BooleanResp> checkHost(
-            @Valid @ModelAttribute @ParameterObject TempCheckHostReq req) {
-
-        TempCheckHostReqDto reqDto = tempAppointmentVoMapper.to(req);
-        boolean isValid = tempAppointmentFacadeService.checkHost(reqDto);
-
-        return ResponseEntity.ok(
-                BooleanResp.of(isValid)
-        );
-    }
-
-    @GetMapping("/join:check")
-    @Operation(summary = "비회원의 약속 참여 여부 조회", description = "비회원의 약속 참여 여부를 조회할 때 사용됩니다. [figma #19,#27]")
-    public ResponseEntity<BooleanResp> checkJoin(
-            @Valid @ModelAttribute @ParameterObject TempCheckJoinReq req) {
-
-        TempCheckJoinReqDto reqDto = tempAppointmentVoMapper.to(req);
-        boolean isValid = tempAppointmentFacadeService.checkJoin(reqDto);
-
-        return ResponseEntity.ok(
-                BooleanResp.of(isValid)
-        );
-    }
-
-    @GetMapping("/user:check")
-    @Operation(summary = "비회원의 약속 포함 여부 조회", description = "비회원의 특정 약속에 포함 여부를 조회할 때 사용됩니다. [figma #27]")
-    public ResponseEntity<BooleanResp> checkUser(
-            @Valid @ModelAttribute @ParameterObject TempCheckUserReq req) {
-
-        TempCheckJoinReqDto reqDto = tempAppointmentVoMapper.to(req);
-        boolean isValid = tempAppointmentFacadeService.checkUser(reqDto);
-
-        return ResponseEntity.ok(
-                BooleanResp.of(isValid)
-        );
-    }
-
-    @GetMapping("/join:valid")
-    @Operation(summary = "비회원의 임시 회원가입 가능 여부 확인 요청", description = "지정 약속에 비회원으로 가입 및 참여가 가능한지 확인합니다. [figma #19]")
-    public ResponseEntity<BooleanResp> validJoin(
-            @Valid @ModelAttribute @ParameterObject CreateTempUserReq req) {
-
-        CreateTempUserReqDto reqDto = tempAppointmentVoMapper.to(req);
-        boolean isValid = tempAppointmentFacadeService.validJoin(reqDto);
-
-        return ResponseEntity.ok(
-                BooleanResp.of(isValid)
-        );
-    }
 }
