+++ src/main/java/com/example/gatherplan/controller/mapper/TempUserVoMapper.java
@@ -2,8 +2,11 @@ package com.example.gatherplan.controller.mapper;
 
 import com.example.gatherplan.appointment.dto.CreateTempUserReqDto;
 import com.example.gatherplan.appointment.dto.TempCheckHostReqDto;
-import com.example.gatherplan.controller.vo.tempappointment.CreateTempUserReq;
-import com.example.gatherplan.controller.vo.tempappointment.TempCheckHostReq;
+import com.example.gatherplan.appointment.dto.TempCheckJoinReqDto;
+import com.example.gatherplan.controller.vo.tempuser.CreateTempUserReq;
+import com.example.gatherplan.controller.vo.tempuser.TempCheckHostReq;
+import com.example.gatherplan.controller.vo.tempuser.TempCheckJoinReq;
+import com.example.gatherplan.controller.vo.tempuser.TempCheckUserReq;
 import org.mapstruct.*;
 
 @Mapper(
@@ -14,6 +17,10 @@ import org.mapstruct.*;
         nullValuePropertyMappingStrategy = NullValuePropertyMappingStrategy.SET_TO_NULL
 )
 public interface TempUserVoMapper {
+
+    TempCheckJoinReqDto to(TempCheckUserReq req);
+
+    TempCheckJoinReqDto to(TempCheckJoinReq req);
 
     CreateTempUserReqDto to(CreateTempUserReq req);
